Fixed the build errors

+++ bagrova.anastasia/T3/Commands.cpp
@@ -13,7 +13,7 @@
 
 using namespace std::placeholders;
 
-bool convertToSize(const std::string& str, int& value) {
+bool convertToSize(const std::string& str, size_t& value) {
     try {
         value = std::stoi(str);
         if (value < 0) {
@@ -28,7 +28,7 @@ bool convertToSize(const std::string& str, int& value) {
 
 bool processArea(const std::string& arg, std::vector<Polygon>& data) {
     // EVEN|ODD|MEAN|<num-of-vertexes:size_t>
-    int num = 0;
+    size_t num = 0;
     if (arg == "EVEN") {
         auto op = [](double init, const Polygon& p) {
             return init + (p.points.size() % 2 == 0 ? p.area() : 0);
@@ -101,7 +101,7 @@ bool processMin(const std::string& arg, std::vector<Polygon>& data) {
 
 bool processCount(const std::string& arg, std::vector<Polygon>& data) {
     // EVEN|ODD|<num-of-vertexes:size_t>
-    int num = 0;
+    size_t num = 0;
 
     auto checkParity = [](int number, int parity) {
         return number % 2 == parity;
